+++ tests/tests/skip-hidden/testdir/progread.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+read ARG1
+read ARG2
+echo $ARG1 $ARG2
+